(#21) Handle multiple rsan IPs and Undef (#23)
* (#21) Handle multiple rsan IPs and Undef

Prior to this commit, when the RSAN IP was empty or had multiple entries
in the PuppetDB query, the catalog would fail to compile. This commit
allows for empty and multiple returned items from the PuppetDB query.

* (#21) Update comments and naming

This commit updates the names of the function that gets IP addresses to
be more inline with its actual job and return type. This commit adds
some comments into the code to make it more clear on what is occuring,
along with some class documentation comments.