Account for non undef postgresql fact (#31)
* Fix an incorrect function name in the exporter

Previously there was an incorrect function name when calling
get_rsan_importer_ips. This commit fixes that incorrect name.

* (#30) Account for non undef postgresql fact

Prior to this commit, the postgresql conditional statements depened on a
fact being undef. In 2019.7.0+, this fact is never undef, so conditional
was no longer viable. This commit adds another conditional to leverage.

* fixup