(maint) Stub out puppet_enterprise in spec tests and add Spec testing CI (#34)
* (maint) Stub out puppet_enterprise in spec tests

Prior to this commit, the module used the real puppet_enterprise module
to do the spec testing. The puppet_enterprise module is an private git
repo, so external testing would not work without authorization. This
commit stubs out the puppet_enterprise module to get the spec tests
working without github authorization.

* (maint) Add spec testing to CI

This commit adds the puppets-epic-show-theatre/action-pdk-test-unit@v1
action to the github workflow.

* (maint) Fix linting errors in spec tests

Prior to this commit, there were some cops that failed during pdk
validate. This commit fixes the issues with code to ensure rubocop does
not show warnings. The commit updates the hash syntax in the spec
testing stubs.

* Change CI workflow to use bundler

Prior to this commit, the github workflow used an action that leveraged
PDK to do the testing and validation. The PDK workflow would fail
installing gem dependencies. This commit updates the CI workflow to use
bundler to execute the validation and spec testing.

* Fix bundle install env variables

Prior to this commit, there were CI failures when installing the gems.
This commit updates the CI files to set the environment variables
correctly to ensure the correct gems are installed.